Make z_score calculation specific to pattern of missing data.
I realized that it would be better and almost easier to just use the observed pattern of missing data to compute the expected mean and variance of the genotype log likelihoods and then use that to compute the z_score.  I have that all good now, and have worked it into self_assign and infer_mixture.