Add a ploidies vector to the param list
A few changes so that rubias can detect haploid markers by every individual having missing data for the second gene copy (column) at the locus.

check_ref_mix now returns a vector of the ploidies of the markers, and that vector makes it into the param list in all the (non-deprecated) functions, so we should be good to go to use those ploidies in the GPROB* macros to compute likelihoods under a haploid model for the haploid markers.