Alter GPROB functions to deal with haploids
No longer is it GPROB_DIP_*.  It is just GPROB, because it handles both diploid and haploid markers.

This compiles and it gives identical results with purely diploid data.  So, now I just need to do some tests to make sure that it is giving me correct results for haploids.